+++ crate/input_ir_rt/src/taffy_to_svg_elements_mapper.rs
@@ -12,6 +12,7 @@ use self::{
     edge_animation_calculator::EdgeAnimationCalculator,
     edge_path_builder_pass_1::EdgePathBuilderPass1,
     edge_path_builder_pass_2::EdgePathBuilderPass2,
+    edge_path_locus_calculator::EdgePathLocusCalculator,
     process_step_heights::ProcessStepsHeight,
     process_step_heights_calculator::ProcessStepHeightsCalculator,
     string_char_replacer::StringCharReplacer,
@@ -29,6 +30,7 @@ mod edge_face_contact_tracker;
 mod edge_model;
 mod edge_path_builder_pass_1;
 mod edge_path_builder_pass_2;
+mod edge_path_locus_calculator;
 mod ortho_protrusion_calculator;
 mod process_step_heights;
 mod process_step_heights_calculator;