+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## unreleased
 
 * Support rendering tooltips. ([#26][#26])
+* Add focus outlines around nodes and edges. ([#27][#27])
+* Update `stroke_style: dashed` to mean `dasharray:4`. ([#27][#27])
 
 [#26]: https://github.com/azriel91/disposition/pull/26
+[#27]: https://github.com/azriel91/disposition/pull/27
 
 
 ## 0.1.0 (2026-04-11)